pyproject.toml
[project]
name = "mykyta-ai-bot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "python-telegram-bot==21.6",
  "requests==2.32.3",
  "python-dotenv==1.0.1",
]
